change license to match the rest of this stuff
Obviously, I won't try to revoke the license on prior versions of
mice. But, from this version onward, I'll be releasing it under the
GNU AGPLv3 (Or later, at your option).
I choose to make this a breaking version increment so that,
if I ever publish mice to Crates.io again, anyone who happens to be
using it will get the chance to review the changes- Cargo won't just
automatically pull in the new, and differently licensed, version.
Not that I think there are very many people using this crate,
anyway. I'm merging it back into mbot because I don't think it's
worth keeping as separate as it is.