make dice rolling benchmark more relevant
In the place I actually need high performance dice rolling, I'm using
a non CSPRNG for the purpose- plotting the PDF for arbitrary dice
expressions. So, we should use the same one here.
(I am considering using a "PRNG" that just returns the same thing
every time, simply to annul the presence of number generation in this
benchmark, since what I really care about is the cost of what we do in
between polls from the RNG in question.)