fix postorder walker and add test
The postorder walker in the tree module was only correct for trees in
which only left hand nodes had children, which happens to be the
variety produced by the dice program parser at present. It now
attempts to descend when moving right, which I forgot to do after
seeing it work for all dice expressions I threw at it.