Add a fuzz target for the stack interpreter, fix bugs
We've been vulnerable to a memory DoS,
and we haven't been handling overflowing arithmetic correctly,
in the stack interpreter.
We now don't trust keep_count for pre-allocation, as it is unlimited.
A good next step might be limiting keep_count values ahead of time,
as it would be nice to not need to worry about this at dice expression
runtime.
We also now perform properly checked arithmetic in KeepHigh terms, as
I had skipped over doing so by mistake before.